Update user functions

+++ src/components/Profile/Profile.tsx
@@ -1,16 +1,23 @@
 import styles from "./Profile.module.css";
 import "../../Variables.css";
 import { useAppDispatch, useAppSelector } from "../../redux/hooks";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { deleteUser, editUser } from "../../redux/api/UserApiCall";
 import { userActions } from "../../redux/slices/User";
-import { useNavigate } from "react-router-dom";
+import { NavLink, useNavigate } from "react-router-dom";
+import { fetchRegisteredHotels } from "../../redux/api/HotelApiCall";
+import { fetchRegisteredTours } from "../../redux/api/TourApiCall";
 
 const Profile = () => {
   const { user } = useAppSelector((state) => state.user);
+  const { registeredHotels } = useAppSelector((state) => state.hotel);
+  const { registeredTours } = useAppSelector((state) => state.tour);
+
   const [firstname, setFirstname] = useState(user.firstname);
   const [lastname, setLastname] = useState(user.lastname);
   const [sidebarOption, setSidebarOption] = useState("account");
+  const [registeredOption, setRegisteredOption] = useState("tours");
+
   const dispatch = useAppDispatch();
   const navigate = useNavigate();
 
@@ -21,17 +28,35 @@ const Profile = () => {
       lastname,
       email: user.email,
       phoneNumber: user.phoneNumber,
+      userHotels: user.userHotels,
+      userTours: user.userTours,
     };
     console.log(userInfo);
     editUser(dispatch, userInfo);
   };
 
   const handleDelete = () => {
-    deleteUser(dispatch);
-    dispatch(userActions.exitFromProfile());
-    localStorage.clear();
-    navigate("/");
+    if (confirm("Are you sure you want to delete your account?")) {
+      deleteUser(dispatch);
+      dispatch(userActions.exitFromProfile());
+      localStorage.clear();
+      navigate("/");
+    }
   };
+
+  const handleRegisteredHotels = () => {
+    setRegisteredOption("hotels");
+    user?.userHotels?.map((hotel) =>
+      fetchRegisteredHotels(dispatch, hotel.hotelId)
+    );
+  };
+
+  const handleRegisteredTours = () => {
+    setRegisteredOption("tours");
+    user.userTours.map((tour) => fetchRegisteredTours(dispatch, tour.tourId));
+  };
+
+  // console.log(registeredHotels);
 
   return (
     <div className={`${styles["user-profile"]} md:flex px-52 py-8`}>
@@ -123,16 +148,67 @@ const Profile = () => {
             <div className="text-3xl font-bold mb-8">
               Recently registered tours and hotels
             </div>
-            <div className={`${styles["registered-options"]}`}>
-              <span>Tours</span>
-              <span>Hotels</span>
+            <div className={`${styles["registered-options"]} mb-4`}>
+              <span
+                className="hover:cursor-pointer hover:text-lime-300"
+                onClick={handleRegisteredTours}
+              >
+                Tours
+              </span>
+              <span
+                className="hover:cursor-pointer hover:text-lime-300"
+                onClick={handleRegisteredHotels}
+              >
+                Hotels
+              </span>
             </div>
-            <div className={`${styles["registered-info"]} flex`}>
-              <div className="flex-auto w-4/12">Tour Image</div>
-              <div className="flex-auto w-8/12">
-                <div>Tour Name</div>
-                <div>Date & Time</div>
-              </div>
+            <div className={styles["registered-info"]}>
+              {registeredOption === "hotels" ? (
+                registeredHotels.length === 0 ? (
+                  <div className="text-xl">No hotels yet.</div>
+                ) : (
+                  registeredHotels?.map((registered) => (
+                    <div className={styles["registered-container"]}>
+                      <div className="flex-auto w-4/12">
+                        <img
+                          className={styles["registered-image"]}
+                          src={registered.hotelImage}
+                          alt=""
+                        />
+                      </div>
+                      <div className="ml-2 flex-auto w-8/12">
+                        <NavLink to="/hotels" className="text-xl font-bold">
+                          {registered.name}
+                        </NavLink>
+                        <div>{registered.city}</div>
+                      </div>
+                    </div>
+                  ))
+                )
+              ) : registeredTours.length === 0 ? (
+                <div className="text-xl">No tours yet.</div>
+              ) : (
+                registeredTours?.map((registered) => (
+                  <div className={styles["registered-container"]}>
+                    <div className="flex-auto w-4/12">
+                      <img
+                        className={styles["registered-image"]}
+                        src={registered.tourImage}
+                        alt=""
+                      />
+                    </div>
+                    <div className="ml-2 flex-auto w-8/12">
+                      <NavLink
+                        to={`/tours/${registered.id}`}
+                        className="text-xl font-bold"
+                      >
+                        {registered.name}
+                      </NavLink>
+                      <div>{registered.country}</div>
+                    </div>
+                  </div>
+                ))
+              )}
             </div>
           </div>
         )}
